requirements.txt
Django>=5.0
elasticsearch>=8.0.0 ; python_version >= '3.8'
requests
redis>=4.0
websockets>=11.0
flask>=2.0
flask-socketio>=5.0
python-socketio>=5.0
eventlet>=0.33
# PostgreSQL driver (install one of these)
psycopg-binary>=3.1